Better handling of raw values vs. human readable values
Change behavior of getting raw data. The human-readable strings are not parsed
until the user requests tags in the callback to `getData()`. The `getData()`
will always save raw values. The default behavior does not change - human
readable values are still returned.

There is an additional boolean parameter on any method returning EXIF tags,
which will translate the EXIF tags to human-readable values when set to false.
When set to true, no translation occurs.

This should allow for sharing the tags with any other program that can then map
the raw tags themselves.

This does not affect IPTC or XMP tags.